Remove provider package.
Provider was being used to attempt to make less requests to Firebase.
The problem with this approach is that when you have a stream of
snapshots being used across multiple screens, the first screen gets the
stream's data immediately but when the stream is used on a subsequent
screen which was created well after the stream, it doesn't get the last
data put into the stream. It only gets any new data put into the stream.
The solution is to request a new stream from Firebase for each screen
that needs it.