Add offline players.
The term "offline players" is kind of a misnomer since the data is
stored on Firebase, but the key difference is the offline player is one
who doesn't have their own device to keep their score on. Instead the
offline players' scores are controlled by everyone else who is not an
offline player.